Moving tests into test/postcss-plugin as requested
